@@ -21,13 +21,16 @@
     "hugo": "cli.js",
     "hugo-extended": "cli.js"
   },
-  "main": "index.js",
+  "type": "module",
+  "exports": "./index.js",
   "dependencies": {
     "chalk": "^4.1.2",
     "decompress": "^4.2.1",
     "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
     "got": "^11.8.2",
-    "sumchecker": "^3.0.1"
+    "sumchecker": "^3.0.1",
+    "tempy": "^2.0.0"
   },
   "devDependencies": {
     "@jakejarvis/eslint-config": "*",
@@ -36,10 +39,10 @@
   },
   "scripts": {
     "postinstall": "node install.js",
-    "test": "eslint \"**/*.js\" && mocha"
+    "test": "eslint . && mocha"
   },
   "engines": {
-    "node": ">=10"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "keywords": [
     "hugo",
@@ -65,15 +68,17 @@
       "@jakejarvis/eslint-config"
     ],
     "parserOptions": {
-      "ecmaVersion": 2018,
+      "ecmaVersion": 2020,
       "sourceType": "module"
     },
     "env": {
-      "node": true,
-      "es6": true
+      "node": true
     },
     "rules": {
       "compat/compat": "off"
-    }
+    },
+    "ignorePatterns": [
+      "vendor/**"
+    ]
   }
 }
