@@ -1,9 +1,8 @@
 #!/usr/bin/env node
 
-"use strict";
+import execa from "execa";
+import hugo from "./index.js";
 
-const execa = require("execa");
-const hugo = require(".");
 const args = process.argv.slice(2);
 
 execa(hugo, args, { stdio: "inherit" });
